requirements.txt
aiohttp==3.12.6
beautifulsoup4==4.13.4
PyYAML==6.0.2
retry==0.9.2
transformers==4.52.4
torch==2.2.2 # Or your specific torch version if different
# For Gemini and LangChain
google-generativeai==0.8.5 # Added by langchain-google-genai usually
langchain==0.3.25
langchain-google-genai==2.0.10
langchain-community # For some tools or utilities if needed later
langchain-core # For pydantic and core types
pydantic # Often a core dependency for LangChain schemas
# For progress bars
tqdm==4.67.1
# For loading .env files
python-dotenv==1.1.0
# For Jina Reader (indirectly via aiohttp)
# For Chroma
chromadb==1.0.12 # Updated ChromaDB version
# For CrewAI workflow orchestration
crewai==0.121.1 